Fixed transaction formatting, added mint and burn cases to activity
