+++ package.json
@@ -31,7 +31,7 @@
     "@fleekhq/browser-rpc": "^2.0.2",
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
-    "@psychedelic/plug-controller": "0.3.1",
+    "@psychedelic/plug-controller": "0.3.2",
     "@psychedelic/plug-inpage-provider": "1.1.15",
     "@reduxjs/toolkit": "^1.6.0",
     "advanced-css-reset": "^1.2.2",
