+++ source/Background/Keyring.js
@@ -1,6 +1,4 @@
-import {
-  E8S_PER_ICP, TOKEN_IMAGES, USD_PER_TC, CYCLES_PER_TC,
-} from '@shared/constants/currencies';
+import { formatAssetBySymbol } from '@shared/constants/currencies';
 import extension from 'extensionizer';
 
 export const NANOS_PER_SECOND = 1_000_000;
@@ -21,27 +19,6 @@ const recursiveParseBigint = (obj) => Object.entries(obj).reduce(
   },
   { ...obj },
 );
-
-const formatAssetBySymbol = (_amount, symbol, icpPrice) => {
-  const amount = parseInt(_amount.toString(), 10);
-  return ({
-    ICP: {
-      amount: amount / E8S_PER_ICP,
-      value: (amount * icpPrice) / E8S_PER_ICP,
-      image: TOKEN_IMAGES.ICP,
-    },
-    XTC: {
-      amount: amount / CYCLES_PER_TC,
-      value: (amount * USD_PER_TC) / CYCLES_PER_TC,
-      image: TOKEN_IMAGES.XTC,
-    },
-    WTC: {
-      amount: amount / CYCLES_PER_TC,
-      value: (amount * USD_PER_TC) / CYCLES_PER_TC,
-    },
-    default: { amount, value: amount },
-  })[symbol || 'default'] || { amount, value: amount };
-};
 
 const formatAssets = (balances, icpPrice) => {
   const mappedAssets = balances.map(({
