+++ source/shared/constants/currencies.js
@@ -40,3 +40,24 @@ export const currencyPropTypes = {
   image: PropTypes.string.isRequired,
 };
 export const CYCLES_PER_TC = 1_000_000_000_000;
+
+export const formatAssetBySymbol = (_amount, symbol, icpPrice) => {
+  const amount = parseInt(_amount.toString(), 10);
+  return ({
+    ICP: {
+      amount: amount / E8S_PER_ICP,
+      value: (amount * icpPrice) / E8S_PER_ICP,
+      image: TOKEN_IMAGES.ICP,
+    },
+    XTC: {
+      amount: amount / CYCLES_PER_TC,
+      value: (amount * USD_PER_TC) / CYCLES_PER_TC,
+      image: TOKEN_IMAGES.XTC,
+    },
+    WTC: {
+      amount: amount / CYCLES_PER_TC,
+      value: (amount * USD_PER_TC) / CYCLES_PER_TC,
+    },
+    default: { amount, value: amount },
+  })[symbol || 'default'] || { amount, value: amount };
+};
